refactor: remove one-time startup data migrations
Remove 5 one-time data migration paths that have already run on production:

- V1DetectionCleanupService (CAS/SeoScraping record cleanup)
- SimilarityHashBackfillService (SimHash column backfill)
- BanCelebrationHashBackfillService (GIF hash column backfill)
- ApiKeysConfig.MigrateLegacyKeys() (dead code, caller deleted in PR #252)
- BackgroundJobConfigService.MigrateOldSettingsFormatAsync() (old Settings dict format)

Also removes DataProcessingConstants (orphaned by SimilarityHashBackfillService),
3 legacy ApiKeysConfig properties, and 19 now-irrelevant tests.

Startup path simplified to: compaction -> EF migrate -> caption seed.

Closes #382
