--- TelegramGroupsAdmin/Constants/DataProcessingConstants.cs
@@ -1,12 +1,0 @@
-namespace TelegramGroupsAdmin.Constants;
-
-/// <summary>
-/// Constants for batch processing and data operations.
-/// </summary>
-public static class DataProcessingConstants
-{
-    /// <summary>
-    /// Batch size for similarity hash backfill operation.
-    /// </summary>
-    public const int SimilarityHashBackfillBatchSize = 500;
-}
